refactor(entities, widgets): remove PopularPosts to widgets
