+++ src/widgets/popular-posts/index.ts
@@ -1,0 +1,1 @@
+export { PopularPosts } from "./ui/popular-posts";
